Use pkg_resources to load calibration data from package files.
CalibrationCurve changed to require a string instead of a file path
(which is better anyway). Since now the user who wants to specify a
calibration curve has to provide the curve canonical name in lower
case (e.g. intcal04, marine04, shcal04).

More flexibility could be added by allowing the user to load a
calibration file in the current directory. It should be easily
implemented.